Added JWT and user protection

+++ backend/controllers/goalController.js
@@ -1,11 +1,13 @@
 const asyncHandler = require('express-async-handler')
+
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 // @desc    Get goals
 // @route   GET /api/goals
 // @access  Private
 const getGoals = asyncHandler(async(req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
     res.status(200).json(goals)
 })
 
@@ -19,7 +21,8 @@ const createGoal = asyncHandler(async(req,res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id,
     })
     res.status(200).json(goal)
 })
@@ -33,6 +36,20 @@ const updateGoal = asyncHandler(async(req, res) => {
     if(!goal){
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    // Check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Ensure logged in user matches goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorised')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true})
@@ -49,7 +66,20 @@ const deleteGoal = asyncHandler(async(req,res) => {
         res.status(400)
         throw new Error('Goal not found')
     }
+    const user = await User.findById(req.user.id)
 
+    // Check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Ensure logged in user matches goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorised')
+    }
+    
     const deleteGoal = await Goal.findByIdAndDelete(req.params.id)
     res.status(200).json(deleteGoal)
 })
